feat: use uv for package management
