--- .codecov.yaml
@@ -1,17 +1,0 @@
-# Based on pydata/xarray
-codecov:
-  require_ci_to_pass: no
-
-coverage:
-  status:
-    project:
-      default:
-        # Require 1% coverage, i.e., always succeed
-        target: 1
-    patch: false
-    changes: false
-
-comment:
-  layout: diff, flags, files
-  behavior: once
-  require_base: no
