--- .cruft.json
@@ -1,29 +1,0 @@
-{
-  "template": "https://github.com/scverse/cookiecutter-scverse",
-  "commit": "75c74be1a7c2ed7e5503510af0c535a3e105ab37",
-  "checkout": null,
-  "context": {
-    "cookiecutter": {
-      "project_name": "CAT",
-      "package_name": "cat",
-      "project_description": "Cluster Alignment Tool (CAT)",
-      "author_full_name": "Alex Valentin Nielsen, Martin Proks, Ala Trusina",
-      "author_email": "yourname@example.com",
-      "github_user": "matq007",
-      "project_repo": "https://github.com/brickmanlab/CAT",
-      "license": "MIT License",
-      "_copy_without_render": [
-        ".github/workflows/build.yaml",
-        ".github/workflows/test.yaml",
-        "docs/_templates/autosummary/**.rst"
-      ],
-      "_render_devdocs": false,
-      "_jinja2_env_vars": {
-        "lstrip_blocks": true,
-        "trim_blocks": true
-      },
-      "_template": "https://github.com/scverse/cookiecutter-scverse"
-    }
-  },
-  "directory": null
-}
