+++ .editorconfig
@@ -11,8 +11,5 @@ insert_final_newline = true
 [*.{yml,yaml}]
 indent_size = 2
 
-[.cruft.json]
-indent_size = 2
-
 [Makefile]
 indent_style = tab
