--- .github/ISSUE_TEMPLATE/feature_request.yml
@@ -1,11 +1,0 @@
-name: Feature request
-description: Propose a new feature for CAT
-labels: enhancement
-body:
-  - type: textarea
-    id: description
-    attributes:
-      label: Description of feature
-      description: Please describe your suggestion for a new feature. It might help to describe a problem or use case, plus any alternatives that you have considered.
-    validations:
-      required: true
