+++ .github/workflows/build.yml
@@ -1,0 +1,28 @@
+name: build
+
+on:
+  push:
+    branches: [master, dev]
+
+jobs:
+  package:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout the repository
+        uses: actions/checkout@main
+
+      - name: Install the latest version of uv
+        uses: astral-sh/setup-uv@v2
+        with:
+          version: "latest"
+          enable-cache: true
+
+      - name: Install Python
+        run: uv python install
+      
+      - name: Install the project
+        run: uv sync --extra dev
+
+      - name: Run ruff
+        run: uvx ruff check src/
