+++ pyproject.toml
@@ -10,130 +10,47 @@ readme = "README.md"
 requires-python = ">=3.10"
 license = {file = "LICENSE"}
 authors = [
-    {name = "Alex Valentin Nielsen"},
-    {name = "Martin Proks"},
-    {name= "Ala Trusina"}
+    {name = "Alex Valentin Nielsen", email = "alexander.nielsen@nbi.ku.dk"},
+    {name = "Martin Proks", email = "martin.proks@sund.ku.dk"},
+    {name= "Ala Trusina", email = "trusina@nbi.ku.dk"}
 ]
 maintainers = [
-    {name = "Martin Proks"},
+    {name = "Martin Proks", email = "martin.proks@sund.ku.dk"},
 ]
-urls.Documentation = "https://CAT.readthedocs.io/"
-urls.Source = "https://github.com/brickmanlab/CAT"
-urls.Home-page = "https://github.com/brickmanlab/CAT"
 dependencies = [
     "anndata",
     "polars",
     "pyarrow",
-    # for debug logging (referenced from the issue template)
-    "session-info"
+    "rich",
+    "rich_argparse",
+    "xlsxwriter"
+]
+classifiers = [
+  "Development Status :: 4 - Beta",
+  "Intended Audience :: Science/Research",
+  "Topic :: Scientific/Engineering :: Artificial Intelligence",
+  "Topic :: Scientific/Engineering :: Bio-Informatics",
+  "License :: OSI Approved :: MIT License",
+  "Programming Language :: Python :: 3.10"
 ]
 
 [project.scripts]
 catcli = "cat.cli:main"
 
 [project.optional-dependencies]
-dev = [
-    "pre-commit",
-    "twine>=4.0.2",
-]
+dev = ["pre-commit"]
 doc = [
-    "docutils>=0.8,!=0.18.*,!=0.19.*",
-    "sphinx>=4",
-    "sphinx-book-theme>=1.0.0",
-    "myst-nb>=1.1.0",
-    "sphinxcontrib-bibtex>=1.0.0",
-    "setuptools",  # Until pybtex >0.23.0 releases: https://bitbucket.org/pybtex-devs/pybtex/issues/169/
-    "sphinx-autodoc-typehints",
-    "sphinxext-opengraph",
-    # For notebooks
-    "ipykernel",
-    "ipython",
-    "sphinx-copybutton",
-]
-test = [
-    "pytest",
-    "coverage",
+  "setuptools",
+  "sphinx>=4.0",
+  "myst-nb>=1.1",
+  "sphinx-book-theme>=1",
+  "sphinxcontrib-bibtex>=1"
 ]
 
-[tool.coverage.run]
-source = ["cat"]
-omit = [
-    "**/test_*.py",
-]
-
-[tool.pytest.ini_options]
-testpaths = ["tests"]
-xfail_strict = true
-addopts = [
-    "--import-mode=importlib",  # allow using test files with same name
-]
+[project.urls]
+Documentation = "https://CAT.readthedocs.io/"
+Source = "https://github.com/brickmanlab/CAT"
+Homepage = "https://github.com/brickmanlab/CAT"
 
 [tool.ruff]
-line-length = 120
-src = ["src"]
-extend-include = ["*.ipynb"]
-
-[tool.ruff.format]
-docstring-code-format = true
-
-[tool.ruff.lint]
-select = [
-    "F",  # Errors detected by Pyflakes
-    "E",  # Error detected by Pycodestyle
-    "W",  # Warning detected by Pycodestyle
-    "I",  # isort
-    "D",  # pydocstyle
-    "B",  # flake8-bugbear
-    "TID",  # flake8-tidy-imports
-    "C4",  # flake8-comprehensions
-    "BLE",  # flake8-blind-except
-    "UP",  # pyupgrade
-    "RUF100",  # Report unused noqa directives
-]
-ignore = [
-    # line too long -> we accept long comment lines; formatter gets rid of long code lines
-    "E501",
-    # Do not assign a lambda expression, use a def -> lambda expression assignments are convenient
-    "E731",
-    # allow I, O, l as variable names -> I is the identity matrix
-    "E741",
-    # Missing docstring in public package
-    "D104",
-    # Missing docstring in public module
-    "D100",
-    # Missing docstring in __init__
-    "D107",
-    # Errors from function calls in argument defaults. These are fine when the result is immutable.
-    "B008",
-    # __magic__ methods are often self-explanatory, allow missing docstrings
-    "D105",
-    # first line should end with a period [Bug: doesn't work with single-line docstrings]
-    "D400",
-    # First line should be in imperative mood; try rephrasing
-    "D401",
-    ## Disable one in each pair of mutually incompatible rules
-    # We don’t want a blank line before a class docstring
-    "D203",
-    # We want docstrings to start immediately after the opening triple quote
-    "D213",
-]
-
-[tool.ruff.lint.pydocstyle]
-convention = "numpy"
-
-[tool.ruff.lint.per-file-ignores]
-"docs/*" = ["I"]
-"tests/*" = ["D"]
-"*/__init__.py" = ["F401"]
-
-[tool.cruft]
-skip = [
-    "tests",
-    "src/**/__init__.py",
-    "src/**/basic.py",
-    "docs/api.md",
-    "docs/changelog.md",
-    "docs/references.bib",
-    "docs/references.md",
-    "docs/notebooks/example.ipynb",
-]
+lint.ignore = ["E731"]
