+++ src/cat/constants.py
@@ -1,0 +1,8 @@
+from importlib.metadata import version
+
+__version__ = version("CAT")
+SIGMA = 1.6
+MIN_N_GENES = 20
+CLUSTER_FIELD = "cat_cluster"
+N_ITERATIONS = 1_000
+DELIMITER = "#"
